feat(mediaLibrary): add media library explorer components and hooks
- Implemented ExplorerBreadcrumbs for navigation within the media library.
- Created ExplorerToolbar for search and filtering options.
- Developed FolderCard and MediaCard for displaying folder and media items.
- Added MediaPreviewModal for previewing media files.
- Introduced MediaUploadPanel for handling file uploads with folder selection.
- Defined types for media library entries and upload configurations.
- Implemented videoThumbnail utility for seeking video thumbnails.
- Created useMediaLibraryExplorer hook for managing media library state and actions.
- Added mediaLibraryPaths utility functions for path manipulation.
